countdown/requirements.txt
flask
gunicorn
firebase-admin
google-cloud-firestore
google-cloud-logging
google-cloud-error-reporting
google-cloud-core